--- lib/speg/error.rb
@@ -1,3 +1,0 @@
-module Speg
-  class Error < StandardError; end
-end
