+++ spec/lib/spec_generator_spec.rb
@@ -1,17 +1,17 @@
 # frozen_string_literal: true
 
-require 'spec_helper'
-require_relative '../../lib/spec_generator'
+# require 'spec_helper'
+# require_relative '../../lib/spec_generator'
 
-RSpec.describe SpecGenerator do
-  let(:path){ ['spec/lib/models'] }
+# RSpec.describe SpecGenerator do
+#   let(:path){ ['spec/lib/models'] }
 
-  it 'generates files' do
-    spec_generator = double()
-    # allow(spec_generator).to receive(:rspec_template).and_return('Rspec template')
-    stub_const('SpecGenerator::PATHS', path)
-    described_class.new.generate_test_files('spec')
+#   it 'generates files' do
+#     spec_generator = double()
+#     # allow(spec_generator).to receive(:rspec_template).and_return('Rspec template')
+#     stub_const('SpecGenerator::PATHS', path)
+#     described_class.new.generate_test_files('spec')
 
-    expect(File.exist?('spec/lib/models/test_spec.rb')).to be_truthy
-  end
-end
+#     expect(File.exist?('spec/lib/models/test_spec.rb')).to be_truthy
+#   end
+# end
