+++ spec/models/dummy_model.rb
@@ -1,0 +1,4 @@
+require 'rails_helper'
+
+RSpec.describe DummyModel do
+end
